Split phpunit tests into unit and integration
