+++ extension.json
@@ -71,7 +71,8 @@
   },
   "load_composer_autoloader": true,
   "TestAutoloadNamespaces": {
-    "MediaWiki\\Extension\\UnifiedExtensionForFemiwiki\\Tests\\Unit\\": "tests/phpunit/unit/"
+    "MediaWiki\\Extension\\UnifiedExtensionForFemiwiki\\Tests\\Unit\\": "tests/phpunit/unit/",
+    "MediaWiki\\Extension\\UnifiedExtensionForFemiwiki\\Tests\\Integration\\": "tests/phpunit/integration/"
   },
   "manifest_version": 2
 }
